+++ src/main/sensors/compass.h
@@ -33,7 +33,7 @@ void compassInit(void);
 void updateCompass(flightDynamicsTrims_t *magZero);
 #endif
 
-extern int16_t magADC[XYZ_AXIS_COUNT];
+extern int32_t magADC[XYZ_AXIS_COUNT];
 
 extern sensor_align_e magAlign;
 extern mag_t mag;
